# backend/tests/fixtures/api_contract.json
{
  "components": {
    "schemas": {
...
      "Candidate": {
        "properties": {
          "depth": {
            "default": 0,
            "title": "Depth",
            "type": "integer"
          },
          "pv": {
            "items": {
              "type": "string"
            },
            "title": "Pv",
            "type": "array"
          },
          "san": {
            "title": "San",
            "type": "string"
          },
          "score": {
            "$ref": "#/components/schemas/Score"
          },
          "uci": {
            "title": "Uci",
            "type": "string"
          }
        },
        "required": [
          "uci",
          "san",
          "score",
          "pv"
        ],
        "title": "Candidate",
        "type": "object"
      },
...
      "ClassificationAudit": {
        "additionalProperties": false,
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "cache_key": {
            "title": "Cache Key",
            "type": "string"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision_id": {
            "title": "Decision Id",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "response": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "cache_key",
          "model",
          "schema_version",
          "prompt_version",
          "status",
          "response",
          "confidence",
          "error",
          "input_tokens",
          "output_tokens",
          "attempts",
          "created_at",
          "provider",
          "version",
          "decision_id"
        ],
        "title": "ClassificationAudit",
        "type": "object"
      },
...
      "CreatedId": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "CreatedId",
        "type": "object"
      },
...
      "Evidence": {
        "additionalProperties": false,
        "properties": {
          "allows_mate": {
            "title": "Allows Mate",
            "type": "boolean"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/Candidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "classifications": {
            "items": {
              "$ref": "#/components/schemas/EvidenceClassification"
            },
            "title": "Classifications",
            "type": "array"
          },
          "facts": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Facts",
            "type": "object"
          },
          "fen": {
            "title": "Fen",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "loss_cp": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loss Cp"
          },
          "mate_lost": {
            "title": "Mate Lost",
            "type": "boolean"
          },
          "played_san": {
            "title": "Played San",
            "type": "string"
          },
          "ply": {
            "title": "Ply",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "fen",
          "ply",
          "played_san",
          "loss_cp",
          "mate_lost",
          "allows_mate",
          "facts",
          "candidates",
          "classifications"
        ],
        "title": "Evidence",
        "type": "object"
      },
      "EvidenceClassification": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "title": "Findings",
            "type": "array"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "skill": {
            "title": "Skill",
            "type": "string"
          }
        },
        "required": [
          "skill",
          "explanation",
          "confidence",
          "run_id",
          "provider",
          "findings"
        ],
        "title": "EvidenceClassification",
        "type": "object"
      },
      "Finding": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "enum": [
              "white",
              "black"
            ],
            "title": "Actor",
            "type": "string"
          },
          "analysis_id": {
            "title": "Analysis Id",
            "type": "string"
          },
          "context_fen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Fen"
          },
          "context_move": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Move"
          },
          "cue": {
            "default": "",
            "title": "Cue",
            "type": "string"
          },
          "direction": {
            "enum": [
              "missed_opportunity",
              "allowed_opponent_tactic"
            ],
            "title": "Direction",
            "type": "string"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "frame_ply": {
            "default": 0,
            "title": "Frame Ply",
            "type": "integer"
          },
          "moves": {
            "items": {
              "type": "string"
            },
            "title": "Moves",
            "type": "array"
          },
          "plies": {
            "items": {
              "type": "integer"
            },
            "minItems": 1,
            "title": "Plies",
            "type": "array"
          },
          "roles": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Roles",
            "type": "object"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "skill_id": {
            "title": "Skill Id",
            "type": "string"
          },
          "squares": {
            "items": {
              "type": "string"
            },
            "title": "Squares",
            "type": "array"
          },
          "verification": {
            "enum": [
              "engine_mate",
              "verified_line",
              "engine_defense"
            ],
            "title": "Verification",
            "type": "string"
          },
          "verification_analysis_ids": {
            "items": {
              "type": "string"
            },
            "title": "Verification Analysis Ids",
            "type": "array"
          }
        },
        "required": [
          "skill_id",
          "rule_id",
          "direction",
          "actor",
          "analysis_id",
          "plies",
          "squares",
          "moves",
          "explanation",
          "verification"
        ],
        "title": "Finding",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "ManualRequest": {
        "properties": {
          "explanation": {
            "default": "",
            "maxLength": 5000,
            "title": "Explanation",
            "type": "string"
          },
          "fen": {
            "maxLength": 200,
            "title": "Fen",
            "type": "string"
          },
          "moves": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Moves",
            "type": "array"
          },
          "orientation": {
            "default": "white",
            "enum": [
              "white",
              "black"
            ],
            "title": "Orientation",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "fen",
          "moves"
        ],
        "title": "ManualRequest",
        "type": "object"
      },
...
      "Rejected": {
        "additionalProperties": false,
        "properties": {
          "rejected": {
            "title": "Rejected",
            "type": "boolean"
          }
        },
        "required": [
          "rejected"
        ],
        "title": "Rejected",
        "type": "object"
      },
...
      "Score": {
        "properties": {
          "kind": {
            "enum": [
              "cp",
              "mate"
            ],
            "title": "Kind",
            "type": "string"
          },
          "mate_given": {
            "default": false,
            "title": "Mate Given",
            "type": "boolean"
          },
          "value": {
            "title": "Value",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "value"
        ],
        "title": "Score",
        "type": "object"
      },
...
      "TeachingAudit": {
        "additionalProperties": false,
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "cache_key": {
            "title": "Cache Key",
            "type": "string"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "response": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "unit_id": {
            "title": "Unit Id",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "cache_key",
          "model",
          "schema_version",
          "prompt_version",
          "status",
          "response",
          "confidence",
          "error",
          "input_tokens",
          "output_tokens",
          "attempts",
          "created_at",
          "unit_id",
          "evidence_ids"
        ],
        "title": "TeachingAudit",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Fieldwork API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/classification-runs/{run_id}": {
      "get": {
        "operationId": "classification_run_api_classification_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassificationAudit"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Classification Run"
      }
    },
    "/api/classification-runs/{run_id}/reject": {
      "post": {
        "operationId": "reject_classification_api_classification_runs__run_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rejected"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Classification"
      }
    },
...
    "/api/evidence/{decision_id}": {
      "get": {
        "operationId": "evidence_api_evidence__decision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "decision_id",
            "required": true,
            "schema": {
              "title": "Decision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evidence"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evidence"
      }
    },
    "/api/exercises/manual": {
      "post": {
        "operationId": "add_manual_api_exercises_manual_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedId"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Manual"
      }
    },
...
    "/api/teaching-runs/{run_id}": {
      "get": {
        "operationId": "teaching_audit_api_teaching_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeachingAudit"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Teaching Audit"
      }
    },
    "/api/teaching-runs/{run_id}/reject": {
      "post": {
        "operationId": "reject_teaching_api_teaching_runs__run_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rejected"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Teaching"
      }
    },
...
  }
}